+++ manim_extensions/__init__.py
@@ -7,19 +7,6 @@
 This package bundles small utility functions and reusable Manim mobjects for
 common geometry, animation, and visualisation tasks.
 
-Examples
---------
-.. manim:: PackageOverviewDocExample
-   :save_last_frame:
-
-   from manim import *
-   from manim_extensions import CircleInt, LabelDot
-
-   class PackageOverviewDocExample(Scene):
-       def construct(self):
-           circle = Circle(radius=1.5, color=BLUE)
-           point = LabelDot("P", [0, 0, 0], label_pos=UP)
-           self.add(circle, point)
 """
 
 from importlib.metadata import version, PackageNotFoundError