+++ manim_extensions/neural_network/__init__.py
@@ -7,16 +7,6 @@
 
 This package provides mobjects for constructing stylised neural network diagrams.
 
-Examples
---------
-.. manim:: NeuralNetworkPackageDocExample
-   :save_last_frame:
-
-   from manim import *
-   from manim_extensions.neural_network import NeuralNetworkMobject
-
-   class NeuralNetworkPackageDocExample(Scene):
-       def construct(self):
 """
 
 from .neural_network import NeuralNetworkMobject