+++ manim_extensions/physics/rigid_mechanics/rigid_mechanics.py
@@ -23,38 +23,6 @@ the specific functions of the space.
         they don't register collisions finely enough. Trying to increase the
         config frame rate might solve the problem.
 
-Examples
---------
-.. manim:: TwoObjectsFalling
-        
-   from manim_extensions.physics import *
-   # use a SpaceScene to utilize all specific rigid-mechanics methods
-   class TwoObjectsFalling(SpaceScene):
-       def construct(self):
-           circle = Circle().shift(UP)
-           circle.set_fill(RED, 1)
-           circle.shift(DOWN + RIGHT)
-
-           rect = Square().shift(UP)
-           rect.rotate(PI / 4)
-           rect.set_fill(PURE_YELLOW_A, 1)
-           rect.shift(UP * 2)
-           rect.scale(0.5)
-
-           ground = Line([-4, -3.5, 0], [4, -3.5, 0])
-           wall1 = Line([-4, -3.5, 0], [-4, 3.5, 0])
-           wall2 = Line([4, -3.5, 0], [4, 3.5, 0])
-           walls = VGroup(ground, wall1, wall2)
-           self.add(walls)
-
-           self.play(
-               DrawBorderThenFill(circle),
-               DrawBorderThenFill(rect),
-           )
-           self.make_rigid_body(rect, circle)  # Mobjects will move with gravity
-           self.make_static_body(walls)  # Mobjects will stay in place
-           self.wait(5)
-           # during wait time, the circle and rect would move according to the simulate updater
 """
 
 from __future__ import annotations